+++ plugins/gs/src/components/scaffolder/ClusterPicker/index.ts
@@ -1,0 +1,3 @@
+export { ClusterPicker } from './ClusterPicker';
+export { ClusterPickerSchema } from './schema';
+export { clusterPickerValidation } from './validation';